+++ CHANGES.md
@@ -19,6 +19,8 @@ Changelog
 ### Changed
 
 - Do not allow code textarea to be dragged to overlap with output sheet.
+- Reject post if title field contains carriage return or line feed character.
+- Reject post if name field contains carriage return or line feed character.
 
 
 ### Fixed